renamed component filenames to pascal case; updated columns filtering for the results page to hide geoid if not present; removed redundant code;
