+++ src/components/results/DataTablePagination.tsx
@@ -6,14 +6,14 @@ import {
   } from "@radix-ui/react-icons"
   import { Table } from "@tanstack/react-table"
    
-  import { Button } from "@/components/ui/button"
+  import { Button } from "@/components/ui/Button"
   import {
     Select,
     SelectContent,
     SelectItem,
     SelectTrigger,
     SelectValue,
-  } from "@/components/ui/select"
+  } from "@/components/ui/Select"
    
   interface DataTablePaginationProps<TData> {
     table: Table<TData>
